# swagger_models.go
package main

import "op-bot/internal/models"

// GitHubUser represents the GitHub user info.
type GitHubUser = models.GitHubUser

// GitHubAppInfo represents the GitHub app installation info.
type GitHubAppInfo = models.GitHubAppInfo

// MeResponse represents the response for /api/github/me.
type MeResponse = models.MeResponse

// ValidateRequest represents the validation request body.
type ValidateRequest = models.ValidateRequest

// ValidationResult represents the validation result.
type ValidationResult = models.ValidationResult